change application structure
