+++ app/app.rb
@@ -1,0 +1,7 @@
+require 'dotenv/load'
+require_relative 'vacancy_pars'
+
+scraper = Scraper.new
+scraper.call
+
+puts "Total Vacancies: #{Vacancy.count}"
