+++ app/db.rb
@@ -1,0 +1,9 @@
+require 'active_record'
+
+ActiveRecord::Base.establish_connection(
+  adapter: 'postgresql',
+  host: 'localhost',
+  database: 'parsing',
+  user: 'bodya',
+  password: 'bodya1'
+)
