+++ app/vacancy.rb
@@ -1,0 +1,2 @@
+class Vacancy < ActiveRecord::Base
+end
